interface-vision/t-044: extract artjob-queue-card.vue from artjob-queue-browser.vue
The per-job <article> block (227 template lines, by far the deepest and most
repeated subtree in a 620-line template) moves into its own component. It's
self-contained: copy/requeue/cancel call the store directly instead of
round-tripping through the parent, and only 'edit' still needs to bubble up
(it opens a modal that lives in the parent). No behavior change.

The magic-number min-height half of the original kaizen complaint (PR #1295)
no longer exists in this file -- already resolved separately before this
pass. Verified: vue-tsc, eslint, prettier all clean; layout-contract holds
with no new violations.
